+++ src/backend/core/tests/test_models_documents.py
@@ -593,6 +593,86 @@ def test_models_documents_get_abilities_preset_role(django_assert_num_queries):
     }
 
 
+@pytest.mark.parametrize(
+    "is_authenticated, is_creator,role,link_reach,link_role,can_destroy",
+    [
+        (True, False, "owner", "restricted", "editor", True),
+        (True, True, "owner", "restricted", "editor", True),
+        (True, False, "owner", "restricted", "reader", True),
+        (True, True, "owner", "restricted", "reader", True),
+        (True, False, "owner", "authenticated", "editor", True),
+        (True, True, "owner", "authenticated", "editor", True),
+        (True, False, "owner", "authenticated", "reader", True),
+        (True, True, "owner", "authenticated", "reader", True),
+        (True, False, "owner", "public", "editor", True),
+        (True, True, "owner", "public", "editor", True),
+        (True, False, "owner", "public", "reader", True),
+        (True, True, "owner", "public", "reader", True),
+        (True, False, "administrator", "restricted", "editor", True),
+        (True, True, "administrator", "restricted", "editor", True),
+        (True, False, "administrator", "restricted", "reader", True),
+        (True, True, "administrator", "restricted", "reader", True),
+        (True, False, "administrator", "authenticated", "editor", True),
+        (True, True, "administrator", "authenticated", "editor", True),
+        (True, False, "administrator", "authenticated", "reader", True),
+        (True, True, "administrator", "authenticated", "reader", True),
+        (True, False, "administrator", "public", "editor", True),
+        (True, True, "administrator", "public", "editor", True),
+        (True, False, "administrator", "public", "reader", True),
+        (True, True, "administrator", "public", "reader", True),
+        (True, False, "editor", "restricted", "editor", False),
+        (True, True, "editor", "restricted", "editor", True),
+        (True, False, "editor", "restricted", "reader", False),
+        (True, True, "editor", "restricted", "reader", True),
+        (True, False, "editor", "authenticated", "editor", False),
+        (True, True, "editor", "authenticated", "editor", True),
+        (True, False, "editor", "authenticated", "reader", False),
+        (True, True, "editor", "authenticated", "reader", True),
+        (True, False, "editor", "public", "editor", False),
+        (True, True, "editor", "public", "editor", True),
+        (True, False, "editor", "public", "reader", False),
+        (True, True, "editor", "public", "reader", True),
+        (True, False, "reader", "restricted", "editor", False),
+        (True, False, "reader", "restricted", "reader", False),
+        (True, False, "reader", "authenticated", "editor", False),
+        (True, True, "reader", "authenticated", "editor", True),
+        (True, False, "reader", "authenticated", "reader", False),
+        (True, False, "reader", "public", "editor", False),
+        (True, True, "reader", "public", "editor", True),
+        (True, False, "reader", "public", "reader", False),
+        (False, False, None, "restricted", "editor", False),
+        (False, False, None, "restricted", "reader", False),
+        (False, False, None, "authenticated", "editor", False),
+        (False, False, None, "authenticated", "reader", False),
+        (False, False, None, "public", "editor", False),
+        (False, False, None, "public", "reader", False),
+    ],
+)
+# pylint: disable=too-many-arguments, too-many-positional-arguments
+def test_models_documents_get_abilities_children_destroy(  # noqa: PLR0913
+    is_authenticated,
+    is_creator,
+    role,
+    link_reach,
+    link_role,
+    can_destroy,
+):
+    """For a sub document, if a user can create children, he can destroy it."""
+    user = factories.UserFactory() if is_authenticated else AnonymousUser()
+    parent = factories.DocumentFactory(link_reach=link_reach, link_role=link_role)
+    document = factories.DocumentFactory(
+        link_reach=link_reach,
+        link_role=link_role,
+        parent=parent,
+        creator=user if is_creator else None,
+    )
+    if is_authenticated:
+        factories.UserDocumentAccessFactory(document=parent, user=user, role=role)
+
+    abilities = document.get_abilities(user)
+    assert abilities["destroy"] is can_destroy
+
+
 @override_settings(AI_ALLOW_REACH_FROM="public")
 @pytest.mark.parametrize(
     "is_authenticated,reach",